When using an architecture that aims to enforce dependency inversion, such as onion, it is very useful for the data layer to look inwardly to the domain and use entities from within.
Before this change, SQLc had the capability to
look to different packages for predefined types
for column types only. This change allows for
functions to receive predefined types as arguments
and return them.

This commit is the first cut of this idea, so it
lacks finesse. One downside is that there is no
capability to use predefined column types in Go
and the new predefined func_name types provided
by this commit.